Patron Singleton y atributos faltantes

+++ agenciauq-client-main/src/main/java/co/edu/uniquindio/agencia/model/Refrigerante.java
@@ -11,6 +11,6 @@ public class Refrigerante extends Producto {
 
     private String codigoAprobacion;
 
-    private double temperaturaRecomendada;
+    private float temperaturaRecomendada;
 
 }
